requirements.txt
streamlit==1.20.0
google-generativeai==0.1.0
pandas==1.5.3
pyyaml==6.0
streamlit-authenticator==0.0.6
fpdf==1.7.2
plotly==5.11.0
streamlit-gsheets==0.0.3
oauth2client==4.1.3
gspread==5.7.1
